Fix ConfigBooleanHotkeyed not (de-)serializing the hotkey part
Also only write the Advanced Keybind Settings to the config if they
are different from the current default settings for a given config.

Note: This fix does change the config data format for that config,
so after this change all of those configs will revert to their
default boolean value.